requirements.txt
scikit-learn>=1.3.0
transformers>=4.33.1
torch>=2.0.0
pandas
Pillow
torchvision
gdown
tabulate
facenet-pytorch
hpsv2
image-reward
clip @ git+https://github.com/openai/CLIP.git
dreamsim